Update example to work with Hercules
Additionally, simplify example structure to have independent files
for each different plugin. This reduces one level of complexity
and makes it easier for the user to read the code and navigate the
components.

The aggregation part has been simplified even more since this example
is the intended to be the first contact for the user with the Processor
and hence it's not needed to show any complex logic.